+++ src/services/metricsservice.h
@@ -9,7 +9,14 @@ class MetricsService : public QObject
 
 public:
     explicit MetricsService(QObject *parent = 0);
-    void sendEvent(const QString &eventCategory = QString(), const QString &eventAction = QString(), const QString &eventLabel = QString(), int eventValue = 0);
+    void sendEventIfEnabled(const QString &eventCategory = QString(),
+                            const QString &eventAction = QString(),
+                            const QString &eventLabel = QString(),
+                            int eventValue = 0);
+    void sendEvent(const QString &eventCategory = QString(),
+                            const QString &eventAction = QString(),
+                            const QString &eventLabel = QString(),
+                            int eventValue = 0);
     void sendHeartbeat();
     void sendAppView(const QString &screenName);
     static MetricsService *instance();